+++ lib/ui/index.dart
@@ -1,4 +1,9 @@
+//Views
 export 'package:sme_app_aluno/ui/views/meus_dados.view.dart';
 export 'package:sme_app_aluno/ui/views/atualizacao_cadastral.view.dart';
 export 'package:sme_app_aluno/ui/views/login.view.dart';
 export 'package:sme_app_aluno/ui/views/meus_dados_editar.view.dart';
+export 'package:sme_app_aluno/ui/views/fluxo_inicial.view.dart';
+
+//Widgets
+export 'package:sme_app_aluno/ui/widgets/ea_botao.widget.dart';
