--- lib/ui/widgets/index.dart
@@ -1,1 +1,0 @@
-export 'package:sme_app_aluno/ui/widgets/ea_botao.widget.dart';
